testing

+++ Modules/Admin/Http/Responses/Lecture/ResumeIndexResponse.php
@@ -4,6 +4,7 @@ namespace Modules\Admin\Http\Responses\Lecture;
 use Illuminate\Contracts\Support\Responsable;
 use Modules\Admin\Entities\Resume;
 use Modules\Admin\Entities\Lecture;
+use Auth;
 use DataTables;
 
 
@@ -42,9 +43,15 @@ class ResumeIndexResponse implements Responsable
             $row['published'] = "<a href='#' onclick='editAct(\"".$resume->id."\",\"".$resume->published."\")'><i class='fa fa-ban' title='edit'></i></a>";
         }
         $row['preview'] = '<a href="/admin/resume/'.$resume->slug.'" id="createbutton" type="button" class=" btn-sm btn-success" style="margin-bottom:5px">'.__("admin::resume.preview").'</a>';
-        $row['action'] = "<a href='/admin/resume/".$resume->id."/edit'><i class='fa fa-pencil-square-o'></i></a>
+        if (Auth::user()->id==$resume->user_id)
+        {
+            $row['action'] = "<a href='/admin/resume/".$resume->id."/edit'><i class='fa fa-pencil-square-o'></i></a>
                         &nbsp;&nbsp;&nbsp;
                     <a href='#' onclick='deleteForm(\"".$resume->id."\")' type='submit'><i class='fa fa-trash'></i></a>";
+        }else{
+
+            $row['action'] = "--";
+        }
         $data[] = $row;
         }
